Refactor penalty handling into dedicated GameCore extension
